requirements.txt
aiobotocore==2.7.0
aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
ansicon==1.89.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.41.2.post2
blessed==1.20.0
boto3==1.28.64
botocore==1.31.64
Brotli @ file:///C:/Windows/Temp/abs_63l7912z0e/croots/recipe/brotli-split_1659616056886/work
certifi==2023.7.22
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
chardet==5.2.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
chitra==0.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
croniter==1.4.1
cryptography @ file:///C:/b/abs_f4do8t8jfs/croot/cryptography_1694444424531/work
cycler==0.12.1
dateutils==0.6.12
debugpy==1.8.0
decorator==5.1.1
deepdiff==6.7.1
docker==6.1.3
docstring-parser==0.15
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
filelock==3.13.1
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
h11==0.14.0
huggingface-hub==0.17.3
hydra-core==1.3.2
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.32.0
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
inquirer==3.1.4
ipykernel==6.26.0
ipython==8.16.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
jinxed==1.2.0
jmespath==1.0.1
jsonargparse==4.27.1
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lazy_loader==0.3
lightning==2.1.2
lightning-api-access==0.0.5
lightning-cloud==0.5.52
lightning-fabric==2.1.2
lightning-utilities==0.9.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/ci/markupsafe_1654508077284/work
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.0.4
nest-asyncio==1.5.8
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
numpy @ file:///C:/b/abs_9fu2cs2527/croot/numpy_and_numpy_base_1695830496596/work/dist/numpy-1.26.0-cp39-cp39-win_amd64.whl#sha256=9df7f35b8a04328650588772e3df359b51f1e83651ab665c98af9dbcf9a088c2
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
ordered-set==4.1.0
packaging==23.2
pandas==2.1.2
parso==0.8.3
pickleshare==0.7.5
Pillow @ file:///C:/b/abs_20ztcm8lgk/croot/pillow_1696580089746/work
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.25.1
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycocotools==2.0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.1.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.1.1
pytz==2023.3.post1
pywin32==306
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
pyzmq==25.1.1
readchar==4.0.5
redis==5.0.1
regex==2023.10.3
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
rich==13.7.0
s3fs==2023.10.0
s3transfer==0.7.0
safetensors==0.4.0
scikit-image==0.22.0
scipy==1.11.3
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
starsessions==1.3.0
supervision==0.16.0
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
tifffile==2023.9.26
timm==0.9.10
tokenizers==0.14.1
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchmetrics==1.2.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
transformers==4.35.0
typeguard==4.1.5
typer==0.9.0
types-python-dateutil==2.8.19.14
typeshed-client==2.4.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.219
urllib3 @ file:///C:/b/abs_9cmlsrm3ys/croot/urllib3_1698257595508/work
uvicorn==0.24.0.post1
wcwidth==0.2.8
websocket-client==1.6.4
websockets==11.0.3
widgetsnbextension==4.0.9
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wrapt==1.16.0
yarl==1.9.2
zipp==3.17.0
